Rewrite: fixed segfault with rewritten URI and "alias".
The "alias" directive cannot be used in the same location where URI
was rewritten.  This has been detected in the "rewrite ... break"
case, but not when the standalone "break" directive was used.

This change also fixes proxy_pass with URI component in a similar
case:

       location  aaa  {
           rewrite ^  xxx yyy;
           break;
           proxy_pass http:  localhost:8080 bbb ;
       }

Previously, the " bbb yyy" would be sent to a backend instead of
" xxx yyy".  And if location's prefix was longer than the rewritten
URI, a segmentation fault might occur.