HTTP 2: lingering close after GOAWAY.
After sending the GOAWAY frame, a connection is now closed using
the lingering close mechanism.

This allows for the reliable delivery of the GOAWAY frames, while
also fixing connection resets observed when http2_max_requests is
reached (ticket  1250), or with graceful shutdown (ticket  1544),
when some additional data from the client is received on a fully
closed connection.

For HTTP 2, the settings lingering_close, lingering_timeout, and
lingering_time are taken from the "server" level.